src/requirements.txt
protobuf >= 2.6.1
flask
gevent
pyzmq
python-dateutil
pymongo
